+++ examples/sqlx_sqlite_migrations/20240101000000_init.sql
@@ -1,0 +1,9 @@
+CREATE TABLE users (
+    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
+    name TEXT NOT NULL
+);
+
+CREATE TABLE op_log (
+    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
+    message TEXT NOT NULL
+);
