+++ examples/sqlx_sqlite_migrations/20240101120000_add_user_version.sql
@@ -1,0 +1,1 @@
+ALTER TABLE users ADD COLUMN version INTEGER NOT NULL DEFAULT 1;
